requirements.txt
discord.py
pillow
pytesseract
opencv-python
requests